refactor submit and sub command for release
